Split impl into generic and specific
